+++ Code/Effects/Challenges/SleepChallenge.cs
@@ -1,4 +1,5 @@
-﻿using CrowdControlMod.ID;
+﻿using CrowdControlMod.CrowdControlService;
+using CrowdControlMod.ID;
 using CrowdControlMod.Utilities;
 
 namespace CrowdControlMod.Effects.Challenges;
@@ -15,6 +16,11 @@ public sealed class SleepChallenge : ChallengeEffect
 
     #region Methods
 
+    protected override CrowdControlResponseStatus OnChallengeStart()
+    {
+        return !GetLocalPlayer().Player.sleeping.isSleeping ? CrowdControlResponseStatus.Success : CrowdControlResponseStatus.Retry;
+    }
+    
     protected override void OnUpdate(float delta)
     {
         if (!GetLocalPlayer().Player.sleeping.isSleeping)