+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Added "Take a Seat" challenge.
 - Added whips to "Give Summon Weapon" effect.
+- Changed some challenges to wait if they would be completed instantly.
 
 ## [2.0.4] - 2023-09-02