+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added "Catch a Critter" challenge.
 - Added whips to "Give Summon Weapon" effect.
 - Changed some challenges to wait if they would be completed instantly.
+- Fixed player morphs not rendering mounts.
 
 ## [2.0.4] - 2023-09-02